Mostly fix broken Quantum Chests
Mostly fixes #129, we still need to address the internal storage values
for automation.

Otherwise we fix the input problem by only allowing the same item on the
input if there is an item in the output slot. We solve the other side of
the input problem by not nulling the internal stack until the output
slot is emptied. this prevents another item from being stored while
there is a different item still in the output slot.