Fix visible count of the quantum chest.
The quantum chest stores items internally but wasn't tracking them well
with the item stack. This causes other mods to not be able to read the
full stored state. As a side effect, this also causes the quantity to
show on the item stack. This fixes it so mods can see the inventory
properly and the GUI doesn't print the count multiple times.